Included list of loaded kernel modules and kernel version information
This can help to infer which hardware matches the loaded modules.

The kernel version is important because the module names and the total
set of available modules change from time to time.

The kernel configuration itself should be available in a different
repository. But the kernel version can be used to match the
corresponding versions.