Replaced runlevel 3 by former runlevel 2 and added extended descriptions
Now runlevel 3 is used as a "minimum services" remote-administration
runlevel, while runlevel 2 serves the same purpose for local logins.

The new runlevel 2 is thus "runlevel S done right", while runlevel S
itself is unusable by definition on systems without a configured "root"
password.

The new definition for runlevel 3 also no longer dictates which services
are required for remote administration, but leaves this definition to
the local system administrator. The usage of the term "LAN" was
misleading anyway, because the old runlevel 2 was used over WAN
connections as well.