Disable -g default option for NTPD
This option is dangerous because it allows a malicious time server to
switch the local machine to any time in the past or future it desires.

This switch should therefore only be used once on startup after bringing
up an Internet connection for the first time, but not with the
permanently running NTP service after that.