Send locale-specific environment variables
Enable SendEnv options for $LC_* and $LANG in the global SSH client
configuration file.

Although this might actually create a problem because there is no way to
override this in the per-user configuration files, it is still possible
for the user to unset the environment variables when invoking ssh.

Also experience has shown that more and more systems only set POSIX
as the locale for SSH logins unless SendEnv is used to override this,
rendering umlauts and UTF-8 features ineffective in the remote shell.

Therefore we too send the locale variables by default now.