Make logrotate update script cron-friendly
The script now checks whether updates are actually necessary and
only uses writes to /tmp (or $TMPDIR) otherwise.

The script now also displays a message in case everything was already
up to date and nothing needed to be done in order to provide some
feedback for the user.

When running the script with stderr redirected to /dev/null, it becomes
cron-friendly and can be run frequently without causing any negative
effects for the lifetime of /etc (in case it is located on flash
storage).

(cherry picked from commit d92f7493a8a25373518386c5c103df915327e730)