Implement a live-patching solution for scripts
The helper shell snippet /etc/libexec/update-patch-helper can now be
sources at the beginning of locally-patched modifications of diverted
upstream scripts.

The helper will detect whether the original script has been modified
(such as by an update), re-apply the patch, and then re-execute the
updated patched script with the same arguments.

The patches for the scripts are kept in "/etc/diversion-patches.d/",
which allows them to remain under version control of THIS branch,
whereas the actually patched files are in /usr or wherever.