Add support for Logitech K830 wireless keyboard
Although the basic functionality of that keyboard works just out of the
box, Xorg customization is required for other functionality UNIX users
are accustomed to.

In particular, that keyboard provides an integrated touchpad with two
mouse buttons, but it does *not* support emulation of the middle mouse
button by default.

Add an Xorg configuration snippet enabling that type of emulation.