Change XDM custom button actions
Instead of actually shutting down or rebooting the machine, display
information how to do achieve this manually. This is less dangerous in
case someone accidentally pressed one of the buttons when the monitor is
switched off.

Also display a digital instead of an analog clock, load the system
default locale, and (currently) assume a German locale.