Recreate Device Mapper nodes on startup
The initramfs does not create the nodes (permanently), so they need to
be re-created before any filesystems can be mounted from such nodes.

There was also a problem that the boot procedure did not mount /dev/pts.
Not sure whether it's a permanent problem or a follow-up of the above
problem. Anyway, create an fstab entry for /dev/pts, so it can be
mounted easier manually if necessary.