New branch with UUID p1qc4h9lc1pndzcfauudk94s4
Initial branch nickname: persistent_live_patcher

Initial check-out location: /etc

Initial version

Provide a framework for maintaining patched copies of original scripts,
where the original scripts may be updated at any time by external means.

The patched copies contain additional shell code at their beginnings
which check for updates in the original scripts every time the copy is
run.

If a change in the original is detected, the copy will be refreshed by
replacing it with the new version of the original and then apply the
patch again.

If applying the patch should fail, the old patched version remains in
place, but a warning for the administrator is written to the syslog.

The framework is especially useful in combination with Debian's
dpkg-divert or similar tools of other distributions. In this case the
original scripts maintained by the distribution can be renamed, and the
patching framework can create copies which have the same name as the
patched files before the diversion.

This branch is intended to be merged into a branch for
version-controlling /etc.