Log to syslog when run as a daemon
Contrary to the decription in the initial commit message, the framework
did not actually write any messages to syslog.

Verify that standard input is a terminal and assume the script to be run
as a daemon otherwise. In the latter case, emit diagnostic messages to
syslog instead of standard error.