Keep last if re-applying the patch should fail
The situation that re-applying the script could fail was not properly
handled by the script so far.

It just removed the patch.

Change this to keep the last working patched version of the script, and
keep logging a warning on every invocation until someone refreshes the
patch so that it can be applied again.