Add cron and logrotate job for log persistency
The cron job will periodically save unsaved portions of BusyBox syslog's
ring buffer to an actual log file.

The logrotation snippet will invoke the cron job with a special
parameter, locking out any access to the log file, so that logrotate can
rotate it without a danger of race conditions.