Fix serveral problems after first initramfs boot
First, it seems the network configuration is still kept in $IP rather
than in $ip. Perhaps this is different on a "real" Debian 9 system?

Second, the script for interactively setting the clock checked the UID
rather than the PID to determine whether it was running from initramfs.

Add /usr/local/* to the PATH. Some of our scripts are actually living
there.

Forgot to add "idn" for password normalization to initramfs. Also, the
Debian BusyBox has configured its "tr" utility not to support POSIX
character classes. Some of my scripts use this feature which is
guaranteed by POSIX, so install the GNU version of the utility instead.

The scripts assumed incorrectly that rngd would remove its PID file once
terminated. This is not true. Use the same algorithm as the init.d
script to startup/shutdown that service.