Lock the serializer fifo before using it
Otherwise users and lockers may not be correlated. The fifo is only
intended to be used by lockers. But users must be serialized, too.
So use a file lock on the fifo for that.

The protocol is now: The user first locks the fifo. Then the user starts
the locker, which locks the actual resource and then read a line from
the fifo. Now the user can use the locked resource. When the user is
done, it sends a line to the fifo. The locker awkens, unlocks the
resource, and exits.