Don't rely on BusyBox symlinks to be installed
Fall back to using the "logread" command only as a last resort. Search
for a BusyBox executable instead, and invoke the "logread" applet on it
if found.

Also try some modified names first before trying "busybox" as the
BusyBox executable, allowing the user to install custom-built versions
of BusyBox which shall be used instead of the default installation.

The idea is that the user renames the custom built BusyBox to one of the
supported names, or (better) creates a symlink in /usr/local/bin with
one of those names which links to the actual executable.