Re-added special permissions
Even though they are obviously host-specific, many of the entries can be
used on several hosts.

Some version with available entries is especially useful when the
permissions in /etc got screwed up and there is no local version of the
special permission file versioned yet. In this case the template version
can be consulted in order to restore at least the most important
permissions.