Make .profile sourcing more tolerant
The .profile initialization files do not always work perfectly. For
instance, if "mesg n" is present but not all file descriptors which
"mesg" assumed to be associated with a terminal device, then it will
generate an intermediate failure return code.

It is therefore not a good idea to source .profile under control of "set
-e". Disable it temporarily.

Anyway, perfection is not necessary in this case: The only reason why
.profile is sourced is to get the root user's normal search $PATH.
Everything else which happens in .profile is of little interest in the
sslcrypt initramfs hook.