Support prioritized overrides for logrotation
The site.d and realm.d subdirectories may contain override snippets with
the same names as those deposited by packages in "/etc/logrotate.d".

In this case, the snippets from the site.d (higher priority) or realm.d
(lower priority) will be copied as the snippets to actually use, rather
than trying to fix the package-provided versions of those snippets.