Update branch with UUID hml46rntp4etzxla7pxzc6rj0
Updated branch nickname: etc_features

Unchanged check-out realm: emgenxx69lwyn5ctlr4nl64ul.local

Updated check-out location: /usr/local/etc/features

This branch used to be a part of /etc. However, it is not really needed
there at runtime, it just needs to be accessed by maintainance
operations.

Therefore, relocate the repository to /usr/local in order to save space
in /etc, which might be restricted.

However, conforming usage requires a some symlink in /etc which points
to the actual check-out directory.

The provided maintenance scripts will make assume this and will
auto-detect the name of the symlink in /etc which points to the
check-out, using that symlink to access the data from the check-out for
synchronizing with the contents of /etc.