tools for yielding ownership in a controlled way
The "give-rpo" script is intended to be used for creating new SCM
repositories in a controlled way using the "sudo"-framework.

To use it, all participating users shall be allowed do assume the role
of the repository master. This role owns the base directory where all
the SCM repositories shall be created.

When a new git repository is to be created, the user who wants to do
this logs in via SSH assuming the repository master's role.

They create the directory where the git repository shall be created, but
do not create the repository yet.

Instead, they invoke the "give-rpo" script via the "sudo" mechanism,
changing the ownership of the new directory to the user who wants to
create a new repository (but only the repository master has the
necessary privileges to do so).

Now the directory belongs to the user who wants to create the new
repositiory, which can do so via "git init --bare" and then push to it.

The repository master, having transferred ownership, cannot do anything
to the new repository now, except renaming it.

In order to delete the repository later, the owner has to delete all the
files in it, leaving only an empty directory. This the repository master
can then remove.