Replace "nbd-server" with "qemu-nbd"
It is the more versatile utility and also supports the "old style" NBD
protocol, which allows to specify an export entirely on the command line
without requiring a configuration file.

This is highly preferable for ad-hoc usages. But 'qemu-nbd" supports the
newer configuration-file based approach also, so no functionality is
lost.

On the other hand, 'qemu-nbd' supports more image types and options than
'nbd-server' does, so it is actually better.