Add compression level recommendation helper
Instead of letting the kernel module compression scripts determine their
"xz" compression level individually, provide a helper script for
centralizing this.

Also change the logic how compression is selected: Instead of being
based on the total amount of memory, let it be based on the amount of
available memory instead.

Also invoke the new helper not at the start of the scripts, but
immediately before compression starts.

This ensures that the memory allocated by temporary files at this point
will also be taken into account.