Consider return codes in ifupdown command options
When using a command option like "pre-up" in an interfaces(5)
interface definition, make sure that the command does not return an
failure code in case of success.

This can easily happen when using "&&" for conditional processing: Make
sure in this case that the command (simple command, command list or
pipeline) does nor return a failure code just because the condition was
not true.

For instance, rather than writing

pre-up my-test && my-conditiona-action

one should write instead

pre-up my-test && my-conditiona-action || :

The "|| :" ensures a "success" return code will still be the result of
the whole command list, even if the conditional "my-test" failed.