Support custom busybox loggers
Introduce the option to specify a custom-build BusyBox version which
overrides the default implementation (but only if actually installed).

Also introduce a custom settings file for the initramfs busybox loggers
which allows to define the beforementioned new option without a need to
modify the existing script.

Rename several add-in conifiguration snippets by adding ".disabled" to
their actual name. This will make such files being ignored at runtime.
However, also add symlinks with the original names which refer to the
".disabled" snippets, making them visible again at runtime.

The idea is that the *.disabled files can always be kept around
unmodified, whether they are needed or not.

And only if they are needed, symlinks without the ".disabled"-suffix are
created which refer to the actual files with the suffix present.

The snippets can therefore be enabled by creating such symlinks, or they
can be disabled by deleting the symlinks.