Add "Quote of the Day" service implementation
It displays a list of passwords which stay the same for some time and
then change. The three latest such passwords are displayed.

The inetd-configuration for this service assumes dictd is also installed
and uses its HOME directory for saving some state. You will have to
create the ~dictd:.local in order for this to work and make it writable
for user "dictd".