Add option for recalculating old QOTD
This is useful if more time has passed since the backlog displayed by
default provides of recent "passwords of the day".

For instance, if you have a file encrypted which such an old password,
display the argument for the new -d option:

$ date -r /path/to/old/encrypted/file +%s

(This assumes your "date" supports the GNU extensions "-r" and "%s".)

Then run the script with the displayed number of seconds since the epoch
like this:

$ /etc/libexec/rfc865server $SECONDS_SINCE_EPOCH

and the "passwords of the day" from the time the file was encrypted will
be displayed.