Add configuration for HUAWEI E3372h-320 modem
This modem is a special case because by default it works as a Web router
instead of a modem.

However, in this default mode, it produces multiple MB of unsolicited
Internet data traffic for its own purposes per day, even if the
connected computer does not transmit a single byte.

Switching the device to an alternate operating mode makes it work like a
normal USB modem again, with the usual access via chat/vdial and pppd.

In order for the alternative device switch to work, however, one has to
complet the setup wizard of the Web router first and Internet access
must have been disabled there.

It is also recommended to switch off all other background activities of
the Web router, such as checking for backdoor updates or new commands
from the Chinese secret service... ehm, I mean, disable checking for
trustworthy firmware updates.