Support xz-compressed kernel modules for RPi
Recent Raspian OS kernels install kernel modules compressed as *.ko.xz
files. Uncompress them before creating an xz-compressed module archive
in /boot.

Also extend this to support other common compressions such as *.ko.gz.
The current assumption ist that there must be an uncompression program
available using the same name as the file extension, except that "un" is
prefixed.

Which means that if a file *.ko.foo is found, than an uncompression
utility "unfoo" will be expected to be installed.