Enhance initramfs dependency declarations
Add "@"-declarations which allow a script to substitute or augment the
functionality of another script.

For instance, if a new script "openssh" emits the dependency
"@dropbear", then it will be treated as a replacement of the real
"dropbear" script dependency-wise. Which means it can act as a drop-in
replacement.

What actually happended is that "openssh" has been made part of the same
dependency group as "dropbear", and all BEFORE/AFTER dependencies only
refer to groups rather than individual scripts. Every script is
automatically made part of a dependency group of the same name, but
other scripts can join this group as well.