Better explain recently added dependency operators
The enhanced dependency operators for initramfs scripts and hooks have
been explained poorly. Explain them (hopefully) somewhat better.

It shall be stated explicitly that "!somename"-references only work for
scripts which have the same name as the group, and are not intended to
be a replacement for group dependencies which remain the primary means
for dependency declarations, even among scripts within the same
dependency group.

The only intended use of "!" is to refer to a script with the identical
name as the group if two scripts are within that group, because refering
to the whole group as a dependency cannot work from inside the group.