Enhance update-patch framework
First of all, it is no longer necessary to create the patch by oneself.
It suffices to create a copy of the original file and modify it as one
sees fit and insert a line for sourcing the update patch helpers
snippet.

When the modified script is run, it will detect that there is no patch
yet and create it from comparing the modified and the unmodified
original version of the script.

Also add an option -u to the update-patch script which will create the
patch immediately rather than waiting that the modified script is
called.

Otherwise, if a patch already exists, it will only be used and never be
recreated. Enforcing the recreation of an exiting patch will only occur
as a consequence of manual invocation of the update-patch script with
the -u option.

This can be useful for "refreshing" a patch which no longer applies
cleanly, i.e. without "fuzz"-warnings.

However, it is not strictly necessary to do so as long as the patch
applied at all, whether cleanly or with fuzz warnings.

Also note that the patch will only be applied once after the original
version has been updated for the last time. Until the next update of the
original, the patched script will then just be run without re-applying
the patch over and over again every time it will be called.