Add 2 potentially useful diversion patches
The first one fixes a big in debdelta which fails when verifying PGP
signatures. It just skips the signature verification, so it is
potentially dangerous. But when verification will fail due to the bug
anyway, there is no harm in applying it.

The second patch fixes a bug in the extfs plugin for the .7z archive
format in avfs. The unmodified plugin script contains a weird workaround
for older versions of p7zip which does not work with current versions of
7za any longer in all situations. The patch removes the workaround,
making the plugin working correcly again in all situations. Note that
the unmodified script works fine in most cases, but not in all.

In particular, I encountered failures if a directory name within the
archive had a period (".") as its last character, such as in "3rd Ed./".
The patch fixed this.

Also note that there is a different version of the same plug in script
which ships with "mc" and does not have to seem that bug either. But as
avfs prefers its own mc extfs extensions, it ignores mc's version of the
script.