Comment out respawn entries without runlevels
It turns out that sysvinit interprets an empty run-level field as "use
in all run-levels" rather than "don't use in any run-level" as expected.

One way to fix this is to set the run-levels field for such entries to
an unused run-level number such as "9".

But commenting the entries out gets rid of the problem easier.