Add a script for simple template replacements
It does not try to apply diffs but rather suggests replacing an entire
/etc file with a template version of vice versa.

The script creates a script containing the suggested instructions which
can be edited by the user. It also creates a diff between the /etc files
and the templates which the user shall take as a basis for his or her
decisions when editing the script.

The user can later run the edited script, actually performing the
scheduled operations.