Fix unintended updates of hard-linked files
The utilities-hook will now remove the target of a copy operation if
that target already exists.

This will break symlinks as well as hardlinks, ensuring no hard-linked
additional files will be changed by the copy command.

Also add a qualifier "~" for utility configuration drop-lists, which
avoids replacing already-existing files and will only copy them if they
do not already exist (as a fallback).