Migrate to new kernel packages on Raspberry Pi
The old kernel-infrastructure packages need to be removed

$ apt-get purge raspberrypi-bootloader raspberrypi-kernel ipxe

because the old kernel was no longer able to verify digital signatures
of certain firmware support files provided by new userspace packages.

To be specific, it was no longer possible to load the "regdomain"
database into the kernel, which is required for Wi-Fi. Therefore, Wi-Fi
could not be used any longer by the old kernel.

In addition, the kernel version in the old packages was severely
outdated. It had not been updated since 2 years ago, and probably never
will.

The new packages ("linux-image*", "linux-header-*") changed the
infrastructure for installing and using them somewhat on the Raspberry
Pi.

For instance, the FAT16 filesystem containing the GPU-Firmware and the
startup configuration (CONFIG.TXT) is no longer mounted on "/", but
rather on "/boot/firmware".

Also, there are two copies now of the kernel- and initrams-images: On
/boot using long filenames which include a version number, and a copy on
/boot/firmware which uses simplified persistent file names without an
embedded kernel version number.

My initramfs creation hook script

rpi-update-bootcopy-all-modules-all-kernels-mmks91b5mk8jm6c6yv3v6h6ee

is therefor no longer necessary, because the name of the initramfs image
file on /boot/firmware does not change any longer on every kernel
update, and can therefore be hard-codes in /boot/firmware/config.txt
rather than being updated by the hook script.

Also, there is no longer a special case for creating archives of the
kernel modules and putting them on /boot/firmware.

Formerly a single package provided multiple kernels, but the kernel
installation hook was only invoked once.

Now the kernels for the different hardware variants of the RPi reside in
separate packages, und the installation hook will be invoked for all of
them.

This works the same as for Debian distros on x86, and thus the same
hook scripts for module archiving can now be used for all platforms.