Add additional utilities for initramfs
For instance, "ne", "ncdu" and "cpipe".

The first is just a replacement for BusyBox' vi, in case more heavy
editing is expected in an initramfs.

"cpipe" has proven to be useful for restraining the write rate to SDXC
cards which otherwise trigger timeout errors in the kernel, which can
lead to write data loss and corrupted filesystems.

"ncdu" is useful for identifying things to delete or move away in case
disk space gets tight.

Also update fstab.manually because of the fact that the default size of
/run has changed from 10 % VM to 20 % VM.

It is also worth noticing that the actual default for /run can now be found
in

/etc/initramfs-tools/initramfs.conf

rather than in /etc/default/tmpfs as it had been before. (The latter
file does not seem to be used by Debian any more.)