Add evidence-burner for standard initramfs
The standard initramfs hook for LMV installs all backups and current LVM
configurations in the initramfs!

While this will not be a problem on unencrypted installation, it
discloses potentially confidental information on (almost - with the
exception of "/boot") full-disk encrypted installations.

Add an additional hook script which removes those files from initramfs,
after the LVM hook did add them.