Rename persistent net rules file
Also add "logsave" utility. Parts of the initramfs framework uses it if
there are problems when mounting or repairing the root fileystem, but
the default hooks do not include it into the initramfs image.

Also fix a problem with the generic crypto mounter dummy script: While
the GNU version of "test" allows to test whether a filesystem is mounted
as read-only, the BusyBox implementation of "test" does not support
this. Use a workaround instead.